feat(payments): 支付渠道——充值+包月订阅（虎皮椒支付宝/微信双渠道、积分批次化）
- 充值：固定4档(¥10/350、¥20/700、¥50/1800、¥100/5000)+自定义金额(≤¥100，
  严格 ¥0.03=1积分，2位小数入账)，永久有效
- 订阅：3档包月限时5折(¥9.99/1000、¥24.99/3000、¥49.99/10000)，30天有效；
  手动续订下周期生效；升级付新档现价全额、旧档未用完积分延期至同期；不降级
- 积分批次化：新表 credit_lots/subscriptions，余额/流水/订单 credits 改
  NUMERIC(20,2)；扣费限时批次优先；退款按 lot_usage 还原原批次；过期惰性
  sweep + 每小时后台任务兜底
- 虎皮椒双渠道凭证（回调双 secret 验签）；GET /payments/catalog 目录接口
- 前端：billing 页重写（续订/升级入口、自定义金额实时换算、渠道选择、下单
  二次确认弹窗）；落地页订阅三档+规则；服务条款第二条重写
- 测试：test_payments 9条 + test_billing_lots 4条，全量 75 passed

+++ src/ai_search/billing/dependencies.py
@@ -6,6 +6,7 @@
 """
 
 import logging
+from decimal import Decimal
 
 from fastapi import Request
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -17,7 +18,7 @@ from .service import InsufficientCreditsError, deduct_credits, refund_credits
 logger = logging.getLogger(__name__)
 
 
-def stamp_request(request: Request, ctx: AuthContext, cost: int, balance: int) -> None:
+def stamp_request(request: Request, ctx: AuthContext, cost: int, balance: Decimal) -> None:
     """把计费/鉴权信息挂到 request.state，供 UsageLog 中间件读取。"""
     request.state.user_id = ctx.user_id
     request.state.api_key_id = ctx.api_key_id
@@ -31,7 +32,7 @@ async def charge_search(
     db: AsyncSession,
     ctx: AuthContext,
     search_depth: str,
-) -> int:
+) -> Decimal:
     """扣费并把信息挂 request.state。返回扣费后余额。
 
     委托纯核 billing.pipeline.charge_credits 做 DB 操作，本层只补
